requirements.txt
hydra-core
matplotlib
seaborn
numpy
scikit-learn
imageio